+++ api/src/reportcreator_api/api_utils/permissions.py
@@ -1,9 +1,10 @@
 from rest_framework.permissions import IsAuthenticated
 
 
-class IsSystemUser(IsAuthenticated):
+class IsAdminOrSystem(IsAuthenticated):
     def has_permission(self, request, view):
-        return super().has_permission(request, view) and request.user.is_system_user
+        return super().has_permission(request, view) and \
+            (request.user.is_system_user or request.user.is_admin)
 
 
 class IsUserManagerOrSuperuserOrSystem(IsAuthenticated):